pyproject.toml
[project]
name = "timetable"
version = "0.1.0"
description = "Time Table for NCN 2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "requests>=2.32.5",
]
